Refactor site forms to separate generator forms
