+++ src/AppBundle/Form/Generator/PostgresType.php
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-namespace AppBundle\Form;
+namespace AppBundle\Form\Generator;
 
-use AppBundle\Entity\Generator\MySQLOptions;
-use AuronConsultingOSS\Docker\Project\ServiceOptions\MySQL;
+use AppBundle\Entity\Generator\PostgresOptions;
+use AuronConsultingOSS\Docker\Project\ServiceOptions\Postgres;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
@@ -26,12 +26,12 @@ use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormInterface;
 
 /**
- * Form for MySQL options.
+ * Form for Postgres options.
  *
- * @package AppBundle\Form
+ * @package AppBundle\Form\Generator
  * @author  Luis A. Pabon Flores
  */
-class MySQLType extends AbstractGeneratorType
+class PostgresType extends AbstractGeneratorType
 {
     /**
      * Builds the form definition.
@@ -42,15 +42,19 @@ class MySQLType extends AbstractGeneratorType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('hasMysql', CheckboxType::class, [
-                'label'    => 'Enable MySQL',
+            ->add('hasPostgres', CheckboxType::class, [
+                'label'    => 'Enable Postgres',
                 'required' => false
             ])
             ->add('version', ChoiceType::class, [
-                'choices'  => array_flip(MySQL::getChoices()),
+                'choices'  => array_flip(Postgres::getChoices()),
                 'expanded' => false,
                 'multiple' => false,
                 'label'    => 'Version'
+            ])
+            ->add('rootUser', TextType::class, [
+                'label' => false,
+                'attr'  => ['placeholder' => 'Root username'],
             ])
             ->add('rootPassword', TextType::class, [
                 'label' => false,
@@ -59,14 +63,6 @@ class MySQLType extends AbstractGeneratorType
             ->add('databaseName', TextType::class, [
                 'label' => false,
                 'attr'  => ['placeholder' => 'Your app\'s database name'],
-            ])
-            ->add('username', TextType::class, [
-                'label' => false,
-                'attr'  => ['placeholder' => 'Your app\'s database username'],
-            ])
-            ->add('password', TextType::class, [
-                'label' => false,
-                'attr'  => ['placeholder' => 'Your app\'s database password'],
             ]);
     }
 
@@ -77,7 +73,7 @@ class MySQLType extends AbstractGeneratorType
      */
     protected function getDataClass()
     {
-        return MySQLOptions::class;
+        return PostgresOptions::class;
     }
 
     /**
@@ -86,12 +82,12 @@ class MySQLType extends AbstractGeneratorType
     protected function getValidationGroups() : callable
     {
         return function (FormInterface $form) {
-            /** @var MySQLOptions $data */
+            /** @var \AppBundle\Entity\Generator\PostgresOptions $data */
             $data   = $form->getData();
             $groups = ['Default'];
 
-            if ($data->hasMysql() === true) {
-                $groups[] = 'mysqlOptions';
+            if ($data->hasPostgres() === true) {
+                $groups[] = 'postgresOptions';
             }
 
             return $groups;
