--- src/AppBundle/Form/VagrantType.php
@@ -1,63 +1,0 @@
-<?php
-/**
- * Copyright 2016 Luis Alberto Pabon Flores
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace AppBundle\Form;
-
-use AuronConsultingOSS\Docker\Project\ServiceOptions\Vagrant;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
-use Symfony\Component\Form\Extension\Core\Type\IntegerType;
-use Symfony\Component\Form\FormBuilderInterface;
-
-/**
- * Vagrant options form.
- *
- * @package AppBundle\Form
- * @author  Luis A. Pabon Flores
- */
-class VagrantType extends AbstractGeneratorType
-{
-    /**
-     * Builds the form definition.
-     *
-     * @param FormBuilderInterface $builder
-     * @param array                $options
-     */
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder
-            ->add('shareType', ChoiceType::class, [
-                'choices'  => array_flip(Vagrant::getChoices()),
-                'expanded' => false,
-                'multiple' => false,
-                'label'    => 'Shared folders filesystem'
-            ])
-            ->add('memory', IntegerType::class, [
-                'label'    => 'RAM available',
-                'required' => true,
-            ]);
-    }
-
-    /**
-     * This should return a string with the FQDN of the entity class associated to this form.
-     *
-     * @return string
-     */
-    protected function getDataClass()
-    {
-        return Vagrant::class;
-    }
-}
